+++ internal/service/repository/gc_test.go
@@ -1,0 +1,89 @@
+package repository
+
+import (
+	"context"
+	"path"
+	"path/filepath"
+	"testing"
+
+	"google.golang.org/grpc/codes"
+
+	"github.com/stretchr/testify/assert"
+	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
+
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+)
+
+func TestGarbageCollectSuccess(t *testing.T) {
+	server := runRepoServer(t)
+	defer server.Stop()
+
+	client := newRepositoryClient(t)
+
+	tests := []struct {
+		req  *pb.GarbageCollectRequest
+		desc string
+	}{
+		{
+			req:  &pb.GarbageCollectRequest{Repository: testRepo, CreateBitmap: false},
+			desc: "without bitmap",
+		},
+		{
+			req:  &pb.GarbageCollectRequest{Repository: testRepo, CreateBitmap: true},
+			desc: "with bitmap",
+		},
+	}
+
+	packPath := path.Join(testhelper.GitlabTestStoragePath(), testRepo.GetRelativePath(), "objects", "pack")
+
+	for _, test := range tests {
+		// Reset mtime to a long while ago since some filesystems don't have sub-second
+		// precision on `mtime`.
+		// Stamp taken from https://golang.org/pkg/time/#pkg-constants
+		testhelper.MustRunCommand(t, nil, "touch", "-t", testTimeString, packPath)
+		t.Log(test.desc)
+		c, err := client.GarbageCollect(context.Background(), test.req)
+		assert.NoError(t, err)
+		assert.NotNil(t, c)
+
+		// Entire `path`-folder gets updated so this is fine :D
+		assertModTimeAfter(t, testTime, packPath)
+
+		bmPath, err := filepath.Glob(path.Join(packPath, "pack-*.bitmap"))
+		if err != nil {
+			t.Fatalf("Error globbing bitmaps: %v", err)
+		}
+		if test.req.GetCreateBitmap() {
+			if len(bmPath) == 0 {
+				t.Errorf("No bitmaps found")
+			}
+		} else {
+			if len(bmPath) != 0 {
+				t.Errorf("Bitmap found: %v", bmPath)
+			}
+		}
+	}
+}
+
+func TestGarbageCollectFailure(t *testing.T) {
+	server := runRepoServer(t)
+	defer server.Stop()
+
+	client := newRepositoryClient(t)
+
+	tests := []struct {
+		repo *pb.Repository
+		code codes.Code
+	}{
+		{repo: nil, code: codes.InvalidArgument},
+		{repo: &pb.Repository{StorageName: "foo"}, code: codes.InvalidArgument},
+		{repo: &pb.Repository{RelativePath: "bar"}, code: codes.InvalidArgument},
+		{repo: &pb.Repository{StorageName: testRepo.GetStorageName(), RelativePath: "bar"}, code: codes.NotFound},
+	}
+
+	for _, test := range tests {
+		_, err := client.GarbageCollect(context.Background(), &pb.GarbageCollectRequest{Repository: test.repo})
+		testhelper.AssertGrpcError(t, err, test.code, "")
+	}
+
+}